requirements.txt
# Core
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic>=2.10.1,<3.0.0
pydantic-settings>=2.6.1,<3.0.0

# Redis for caching & sessions
redis==5.0.1
hiredis==2.2.3

# Logging
python-json-logger==2.0.7

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2